Rename SA1652 to SA0001 and improve messaging
Fixes #2030
