+++ StyleCop.Analyzers/StyleCop.Analyzers/DocumentationRules/DocumentationResources.resx
@@ -285,15 +285,6 @@
   <data name="SA1651CodeFix" xml:space="preserve">
     <value>Finalize placeholder text</value>
   </data>
-  <data name="SA1652Description" xml:space="preserve">
-    <value>The project is currently configured to not parse XML documentation comments; all diagnostics for documentation comments are disabled. Enable XML documentation output to enable these diagnostics.</value>
-  </data>
-  <data name="SA1652MessageFormat" xml:space="preserve">
-    <value>Enable XML documentation output</value>
-  </data>
-  <data name="SA1652Title" xml:space="preserve">
-    <value>Enable XML documentation output</value>
-  </data>
   <data name="StartingTextGets" xml:space="preserve">
     <value>Gets</value>
   </data>
