+++ StyleCop.Analyzers/StyleCop.Analyzers/SpecialRules/SpecialResources.resx
@@ -133,4 +133,13 @@ Note that some situations are not affected by the bug:
   <data name="SA0000Title" xml:space="preserve">
     <value>Workaround incomplete diagnostics in Visual Studio 2015 Update 1</value>
   </data>
+  <data name="SA0001Description" xml:space="preserve">
+    <value>XML comment analysis can only be performed when the project is configured to parse documentation comments. To enable this functionality, update the project to produce an XML documentation file as part of the build.</value>
+  </data>
+  <data name="SA0001MessageFormat" xml:space="preserve">
+    <value>XML comment analysis is disabled due to project configuration</value>
+  </data>
+  <data name="SA0001Title" xml:space="preserve">
+    <value>XML comment analysis disabled</value>
+  </data>
 </root>
