+++ StyleCopAnalyzers.sln
@@ -38,6 +38,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "documentation", "documentat
 		documentation\EnableConfiguration.md = documentation\EnableConfiguration.md
 		documentation\KnownChanges.md = documentation\KnownChanges.md
 		documentation\SA0000Roslyn7446Workaround.md = documentation\SA0000Roslyn7446Workaround.md
+		documentation\SA0001.md = documentation\SA0001.md
 		documentation\SA1000.md = documentation\SA1000.md
 		documentation\SA1001.md = documentation\SA1001.md
 		documentation\SA1002.md = documentation\SA1002.md
